Refactor skillwallet routing
